renamed topBanner file name
