+++ client/src/components/topBanner/index.js
